--- Application/DTO/BaseEntityDto.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.DTO
-{
-    public class BaseEntityDto
-    {
-        public int Id { get; set; }
-    }
-}
